// frontend/src/api/generated/zod/alunos/alunos.ts
export const AlunosControllerCreateBody = zod.object({
  "nome": zod.string().min(1),
  "matricula": zod.string().min(1).describe('Número de matrícula, também usado como login')
})