// frontend/src/api/generated/models/alunoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */

export interface AlunoDto {
  id: string;
  userId: string;
  nome: string;
  matricula: string;
  /** @nullable */
  turmaId: string | null;
  /** @nullable */
  fotoUrl: string | null;
}